requirements.txt
django-cors-headers==4.3.0
django-filter==23.3
django-phone-field==1.8.1
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
drf-yasg==1.21.7
ecdsa==0.18.0
pymongo==3.12.3
psycopg2-binary==2.9.9
mongoengine==0.27.0
djongo==1.3.6
packaging==23.2
setuptools==68.2.2
twilio==9.2.1
google-cloud-storage==2.17.0
cloudinary==1.40.0
Pillow==10.1.0
django-cloudinary-storage==0.3.0
boto==2.49.0